requirements.txt
# Trading Bot — production dependencies
numpy>=1.24.0
pandas>=2.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
requests>=2.28.0
python-binance>=1.0.19
pytest>=7.0.0
